requirements.txt
flask>=3.0.0
anthropic>=0.45.0
apscheduler>=3.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
python-pptx>=0.6.23
openpyxl>=3.1.0
